feat: delete contact activity indicator

+++ src/screens/ContactDetail/index.js
@@ -1,7 +1,8 @@
 import { useNavigation, useRoute } from '@react-navigation/native'
 import { Icon } from '@rneui/themed'
-import React, { useContext, useEffect, useReducer } from 'react'
-import { TouchableOpacity, View } from 'react-native'
+import React, { useContext, useEffect } from 'react'
+import { ActivityIndicator, Alert, TouchableOpacity, View } from 'react-native'
+import colors from '../../assets/theme/colors'
 import ContactDetailsComponent from '../../components/ContactDetails'
 import { CONTACT_LIST } from '../../constants/routeNames'
 import deleteContact from '../../context/actions/contacts/deleteContact'
@@ -16,7 +17,7 @@ const ContactDetail = () => {
   const {
     contactsDispatch,
     contactsState: {
-      getContacts: { data, loading },
+      deleteContact: { loading },
     },
   } = useContext(GlobalContext)
 
@@ -29,18 +30,38 @@ const ContactDetail = () => {
             <View>
               <TouchableOpacity
                 onPress={() => {
-                  deleteContact(item.id)(contactsDispatch)(() => {
-                    navigate(CONTACT_LIST, { successDelete: true })
-                  })
+                  Alert.alert(
+                    'Delete contact',
+                    `You are going to delete ${item.first_name} ${item.last_name}. Are you sure?`,
+                    [
+                      {
+                        text: 'Cancel',
+                        onPress: () => {},
+                      },
+                      {
+                        text: 'Ok',
+                        onPress: () => {
+                          deleteContact(item.id)(contactsDispatch)(() => {
+                            navigate(CONTACT_LIST, { successDelete: true })
+                          })
+                        },
+                      },
+                    ]
+                  )
                 }}>
-                <Icon type="ionicon" name={'trash'} color={'darkgrey'} />
+                {!loading && (
+                  <Icon type="ionicon" name={'trash'} color={'darkgrey'} />
+                )}
+                {loading && (
+                  <ActivityIndicator size={20} color={colors.secondary} />
+                )}
               </TouchableOpacity>
             </View>
           )
         },
       })
     }
-  }, [item, data.length])
+  }, [item, loading])
 
   return <ContactDetailsComponent contact={item} />
 }
